+++ areas/cloudarchitect/src/AzureMcp.CloudArchitect/Commands/BaseCloudArchitectCommand.cs
@@ -1,0 +1,15 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+using System.Diagnostics.CodeAnalysis;
+using AzureMcp.CloudArchitect.Options;
+using AzureMcp.Core.Commands;
+
+namespace AzureMcp.CloudArchitect.Commands;
+
+public abstract class BaseCloudArchitectCommand<
+    [DynamicallyAccessedMembers(TrimAnnotations.CommandAnnotations)] T>
+    : GlobalCommand<T>
+    where T : BaseCloudArchitectOptions, new()
+{
+}